DOC: Fix examples and docs
