+++ README.md
@@ -73,71 +73,174 @@ These are all directly equivalent to the helm commands.
 The real power of avionix comes from it being object oriented. The one thing that
  helm is missing that avionix has is inheritance.
 ```python
-from typing import Optional
+from typing import List, Optional
 
-from avionix import ObjectMeta
-from avionix.kubernetes_objects.core import Service, ServicePort, ServiceSpec
+from my_package import (
+    AirflowLogVolumeGroup,
+    AirflowOptions,
+    RedisOptions,
+    SqlOptions,
+    ValueOrchestrator,
+    AirflowDagVolumeGroup,
+    ExternalStorageVolumeGroup,
+)
+
+from avionix.kubernetes_objects.core import (
+    Container,
+    ContainerPort,
+    EnvFromSource,
+    EnvVar,
+    Probe,
+    SecretEnvSource,
+    HTTPGetAction,
+)
 
 
-class MyBaseService(Service):
+class AvionixAirflowProbe(Probe):
+    def __init__(self, path: str, port: int, host: Optional[str] = None):
+        super().__init__(http_get=HTTPGetAction(path=path, port=port, host=host))
+
+
+class CoreEnvVar(EnvVar):
+    def __init__(self, name: str, value):
+        super().__init__("AIRFLOW__CORE__" + name, value)
+
+
+class AirflowContainer(Container):
     def __init__(
         self,
         name: str,
-        port: int,
-        target_port: int,
-        node_port: int,
-        selector_labels: dict,
-        external_name: Optional[str] = None,
-        port_name: Optional[str] = None,
-        protocol: Optional[str] = None,
-        test_mode: bool = False,
+        sql_options: SqlOptions,
+        redis_options: RedisOptions,
+        airflow_options: AirflowOptions,
+        ports: Optional[List[ContainerPort]] = None,
+        readiness_probe: Optional[Probe] = None,
     ):
+        values = ValueOrchestrator()
+        self._sql_options = sql_options
+        self._redis_options = redis_options
+        self._airflow_options = airflow_options
         super().__init__(
-            ObjectMeta(name=name),
-            ServiceSpec(
-                [
-                    ServicePort(
-                        name=port_name,
-                        port=port,
-                        target_port=target_port,
-                        node_port=node_port if test_mode else None,
-                        protocol=protocol,
-                    )
-                ],
-                selector=selector_labels,
-                external_name=external_name,
-                type="NodePort" if test_mode else "ClusterIP",
-                external_traffic_policy="Local" if test_mode else None,
+            name=name,
+            args=[name],
+            image="airflow-image",
+            image_pull_policy="Never",
+            env=self._get_environment(),
+            env_from=[
+                EnvFromSource(
+                    None, None, SecretEnvSource(values.secret_name, optional=False)
+                )
+            ],
+            ports=ports,
+            volume_mounts=self._get_volume_mounts(),
+            readiness_probe=readiness_probe,
+            command=["/entrypoint.sh"],
+        )
+
+    def _get_volume_mounts(self):
+        return [
+            AirflowLogVolumeGroup(self._airflow_options).volume_mount,
+            AirflowDagVolumeGroup(self._airflow_options).volume_mount,
+            ExternalStorageVolumeGroup(self._airflow_options).volume_mount,
+        ]
+
+    def _get_environment(self):
+        env = (
+            self._get_kubernetes_env()
+            + self._get_airflow_env()
+            + self._airflow_options.extra_env_vars
+        )
+        return env
+
+    def _get_airflow_env(self):
+        return [
+            CoreEnvVar("EXECUTOR", self._airflow_options.core_executor),
+            CoreEnvVar("DEFAULT_TIMEZONE", self._airflow_options.default_time_zone,),
+            CoreEnvVar("LOAD_DEFAULT_CONNECTIONS", "False"),
+            CoreEnvVar("LOAD_EXAMPLES", "False"),
+            CoreEnvVar(
+                "DAGS_ARE_PAUSED_AT_CREATION",
+                str(self._airflow_options.dags_paused_at_creation),
             ),
-        )
+        ]
+
+    def _get_kubernetes_env(self):
+        return [
+            EnvVar("AIRFLOW__KUBERNETES__NAMESPACE", self._airflow_options.namespace),
+            EnvVar(
+                "AIRFLOW__KUBERNETES__DAGS_VOLUME_CLAIM",
+                AirflowDagVolumeGroup(
+                    self._airflow_options
+                ).persistent_volume_claim.metadata.name,
+            ),
+            EnvVar(
+                "AIRFLOW__KUBERNETES__LOGS_VOLUME_CLAIM",
+                AirflowLogVolumeGroup(
+                    self._airflow_options
+                ).persistent_volume_claim.metadata.name,
+            ),
+            EnvVar(
+                "AIRFLOW__KUBERNETES__WORKER_CONTAINER_REPOSITORY",
+                self._airflow_options.worker_image,
+            ),
+            EnvVar(
+                "AIRFLOW__KUBERNETES__WORKER_CONTAINER_IMAGE_PULL_POLICY",
+                "IfNotPresent",
+            ),
+            EnvVar(
+                "AIRFLOW__KUBERNETES__WORKER_CONTAINER_TAG",
+                self._airflow_options.worker_image_tag,
+            ),
+        ]
 
 
-class ChildService1(MyBaseService):
+class WebserverUI(AirflowContainer):
     def __init__(
         self,
-        name: str,
-        port: int,
-        target_port: int,
-        node_port: int,
-        selector_labels: dict,
-        external_name: Optional[str],
-        port_name: Optional[str],
-        protocol: Optional[str],
-        test_mode: bool = False,
+        sql_options: SqlOptions,
+        redis_options: RedisOptions,
+        airflow_options: AirflowOptions,
     ):
         super().__init__(
-            name,
-            port,
-            target_port,
-            target_port,
-            node_port,
-            selector_labels,
-            external_name,
-            port_name,
-            protocol,
-            test_mode,
+            "webserver",
+            sql_options,
+            redis_options,
+            airflow_options,
+            ports=[ContainerPort(8080, host_port=8080)],
+            readiness_probe=AvionixAirflowProbe("/airflow", 8080, "0.0.0.0"),
         )
 
+
+class Scheduler(AirflowContainer):
+    def __init__(
+        self,
+        sql_options: SqlOptions,
+        redis_options: RedisOptions,
+        airflow_options: AirflowOptions,
+    ):
+        super().__init__(
+            "scheduler",
+            sql_options,
+            redis_options,
+            airflow_options,
+            [ContainerPort(8125, host_port=8125)],
+        )
+
+
+class FlowerUI(AirflowContainer):
+    def __init__(
+        self,
+        sql_options: SqlOptions,
+        redis_options: RedisOptions,
+        airflow_options: AirflowOptions,
+    ):
+        super().__init__(
+            "flower",
+            sql_options,
+            redis_options,
+            airflow_options,
+            readiness_probe=AvionixAirflowProbe("/flower/", 5555),
+        )
 ```
 
 # Development Environmnent
@@ -157,5 +260,26 @@ minikube start --driver=docker
 
 ## Warnings
 
-Note that you currently cannot use public variables if subclassing one of the
+Note that you currently cannot use public instance variables if subclassing one of the
  kubernetes components.
+ 
+For example in the above inhertiance example code, this would break the helm output,
+
+```python
+class WebserverUI(AirflowContainer):
+    def __init__(
+        self,
+        sql_options: SqlOptions,
+        redis_options: RedisOptions,
+        airflow_options: AirflowOptions,
+    ):
+        self.my_personal_variable = "I'm breaking helm!"
+        super().__init__(
+            "webserver",
+            sql_options,
+            redis_options,
+            airflow_options,
+            ports=[ContainerPort(8080, host_port=8080)],
+            readiness_probe=AvionixAirflowProbe("/airflow", 8080, "0.0.0.0"),
+        )
+```
